requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
Brotli==1.0.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter_server==2.14.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
lightning==2.2.5
lightning-utilities==0.11.3.post0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.6
mistune==2.0.4
monai==1.3.2
mpmath==1.3.0
multidict==6.0.5
#nb-conda==2.2.1
#nb-conda-kernels==2.3.1
#nbclassic==1.1.0
#nbclient==0.8.0
#nbconvert==7.10.0
#nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
notebook==6.5.7
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydicom==2.4.4
Pygments==2.15.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==2.9.0
setproctitle==1.3.3
setuptools==69.5.1
SimpleITK==2.3.1
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
terminado==0.17.1
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.10
xlrd==2.0.1
y-py==0.5.9
yarl==1.9.4
ypy-websocket==0.8.2
zipp==3.17.0
mislight @ git+https://github.com/jwc-rad/MISLight.git@606d240cb9167f4146744e601e3aebf9f412551b